refactor: correct the folder structure for Footer
