+++ src/parts/Footer/index.tsx
